+++ source/extensions/omni.isaac.ros2_bridge/nodes/OgnROS2SubscribeTwist.cpp
@@ -7,23 +7,23 @@
 // license agreement from NVIDIA CORPORATION is strictly prohibited.
 //
 
-#include "rosgraph_msgs/msg/clock.hpp"
+#include "geometry_msgs/msg/twist.hpp"
 
 #include <omni/isaac/ros/Ros2Node.h>
 
-#include <OgnROS2SubscribeClockDatabase.h>
+#include <OgnROS2SubscribeTwistDatabase.h>
 
-class OgnROS2SubscribeClock : public Ros2Node
+class OgnROS2SubscribeTwist : public Ros2Node
 {
 public:
     // static void initialize(const GraphContextObj& contextObj, const NodeObj& nodeObj)
     // {
-    //     auto& state = OgnROS2SubscribeClockDatabase::sInternalState<OgnROS2SubscribeClock>(nodeObj);
+    //     auto& state = OgnROS2SubscribeTwistDatabase::sInternalState<OgnROS2SubscribeTwist>(nodeObj);
     // }
 
-    static bool compute(OgnROS2SubscribeClockDatabase& db)
+    static bool compute(OgnROS2SubscribeTwistDatabase& db)
     {
-        auto& state = db.internalState<OgnROS2SubscribeClock>();
+        auto& state = db.internalState<OgnROS2SubscribeTwist>();
         // spin once calls reset automatically if it was not successful
         const auto& nodeObj = db.abi_node();
         if (!state.spinOnce(
@@ -31,6 +31,7 @@ public:
         {
             return false;
         }
+
         // Subscriber was not valid, create a new one
         if (!state.mSubscriber)
         {
@@ -40,10 +41,10 @@ public:
             {
                 return false;
             }
-            state.mCallback = [&state, &db](const rosgraph_msgs::msg::Clock::SharedPtr msg)
+            state.mCallback = [&state, &db](const geometry_msgs::msg::Twist::SharedPtr msg)
             { state.subCallback(msg, db); };
 
-            state.mSubscriber = state.mNodeHandle->create_subscription<rosgraph_msgs::msg::Clock>(
+            state.mSubscriber = state.mNodeHandle->create_subscription<geometry_msgs::msg::Twist>(
                 fullTopicName, db.inputs.queueSize(), state.mCallback);
             return true;
         }
@@ -53,7 +54,7 @@ public:
 
     static void release(const NodeObj& nodeObj)
     {
-        auto& state = OgnROS2SubscribeClockDatabase::sInternalState<OgnROS2SubscribeClock>(nodeObj);
+        auto& state = OgnROS2SubscribeTwistDatabase::sInternalState<OgnROS2SubscribeTwist>(nodeObj);
         state.reset();
     }
 
@@ -70,16 +71,27 @@ public:
         Ros2Node::reset();
     }
 
-    void subCallback(const rosgraph_msgs::msg::Clock::SharedPtr msg, OgnROS2SubscribeClockDatabase& db)
+    void subCallback(const geometry_msgs::msg::Twist::SharedPtr& msg, OgnROS2SubscribeTwistDatabase& db)
     {
-        db.outputs.timeStamp() = rclcpp::Time(msg->clock).seconds();
+        auto& linVel = db.outputs.linearVelocity();
+
+        linVel[0] = msg->linear.x;
+        linVel[1] = msg->linear.y;
+        linVel[2] = msg->linear.z;
+
+        auto& angVel = db.outputs.angularVelocity();
+
+        angVel[0] = msg->angular.x;
+        angVel[1] = msg->angular.y;
+        angVel[2] = msg->angular.z;
+
         db.outputs.execOut() = kExecutionAttributeStateEnabled;
     }
 
 
 private:
-    std::shared_ptr<rclcpp::Subscription<rosgraph_msgs::msg::Clock>> mSubscriber = nullptr;
-    std::function<void(const rosgraph_msgs::msg::Clock::SharedPtr)> mCallback;
+    std::shared_ptr<rclcpp::Subscription<geometry_msgs::msg::Twist>> mSubscriber = nullptr;
+    std::function<void(const geometry_msgs::msg::Twist::SharedPtr)> mCallback;
 };
 
 REGISTER_OGN_NODE()